zq update through "add support for three basic joins" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2210, authored by @mccanne,
has been merged.

add support for three basic joins

This commit extends the current join syntax to support "inner join",
"left join", and "right join".  The default "join" is now an
inner join (to follow SQL convention) so this will break any scripts
or workflow out there that assumes the left-outer behavior of the
previous "join" command.