zq update through "Aurora helm" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2161, authored by @marktwallace,
has been merged.

Aurora helm

This PR consists of several sub-projects that all needed to work before Aurora could be used and tested.

(1) An Aurora cluster has been created. The instructions in aurora.md can be used to reproduce the configuration, but this only needs to be done once per AWS account, so we should not do it again until we start a Prod account.

(2) The Aurora cluster was created in a separate VPC from EKS. This will allow us to reconfigure EKS or create additional clusters without needing to change our Aurora cluster. We set up VPC peering from the EKS VPC to the Aurora VPC. Peering is a tricky to test and time-consuming setup process, so we created a script (peering.sh) so the key steps can be reproduced when we need to do it again for our Prod cluster. As per best practices for security, the Aurora cluster is not publicly accessible. It can only be accessed from within our EKS cluster.

(3) The Helm K8s deployment scripts needed to be restructured to allow Postgres (and Redis) to be optional containers. We use Helm "tags" for the optional features.

(4) Makefile targets needed to be adjusted and a new target created for deployments that use Aurora.

(5) All development and test activities will share a single Aurora cluster. Like EKS, Aurora can become very expensive if each developer needs their own cluster, so we developed a set of conventions to share the Aurora cluster that corresponded to the way we share the EKS test cluster. Each developer or test environment will use a separate Aurora database within the cluster. This provides ideal isolation for testing, which is a good match for the namespace isolation we use in our EKS cluster. There are instructions in aurora.md for how to create Aurora test users and databases using kubectl and psql. It is a manual process, and I will be happy to do it for developers that do not want to deep dive into kubectl and psql.

(6) The cd.yaml tests have been updated to add tests for Aurora. The master deploy now runs on Aurora, which means that environment will be stateful and use db migrations to maintain consistent state as we develop. The master deploy now runs the same tests we run in the test namespace, but runs them with a migrated Aurora db rather than a new, empty database. As we add more tests, the tests for master may diverge from the smoke tests that run in a temporary namespace, because of the need to support migrations.

(7) After upgrading the EKS cluster to K8s 1.19, we started having failures in the automated test in cd.yaml. There were multiple causes for this:
  (a) The new node group created lacked permission to access S3. (It was hard to find this omission.) The eks-setup.md instructions have been updated to add procedures for upgrading the cluster and modifying node groups so we can avoid this in the future.
  (b) Linkerd did not work properly with EKS 1.19 -- the problem was subtle, and showed up when the cd.yaml test would hang indefinitely when trying to remove a temporary namespace (e.g. ci-588). We fixed this by removing Linkerd, which I do not think we will need going forward because we plan on using AWS AMP for storing metrics, and the native metrics for K8s 1.19 are thorough enough that Linkerd does not add any metrics we need.
  (c) In addition, the pattern for temporary namespaces has been changed in the cd.yaml tests. The new tests use no more than 10 temp namespaces and recycle the old ones. I think this is more prudent and will avoid possible future problems with namespace deletion.