zq update through "Move pcap test helpers to handlers_pcap_test.go" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2238, authored by @nwt,
has been merged.

Move pcap test helpers to handlers_pcap_test.go

It's a better home for them since it's the only file in which they're used. This also prevents tools from flagging them as dead code since they're now defined—and not just used—under the `pcapingest` build constrait.