zq update through "add zson_parse function" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2242, authored by @alfred-landrum,
has been merged.

add zson_parse function

We'll want a way to make metadata available to the shapers that run for an intake. One way we could do that is by creating a ZSON definition of the metadata record, then prepending the `const` definition of that record to the shaper's source. That requires a way to parse ZSON inside of Z, so I put this up to get feedback on the overall approach.