zq update through "use ztest for expr/expr_test.go" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2251, authored by @mccanne,
has been merged.

use ztest for expr/expr_test.go

This commit changes expr_test.go to use ztest instead of reaching
into the compiler library to compile a standalone expression.
This arrangement is easier to maintain and debug failing tests.

This is needed by a subsequent PR that is changing the compiler
internals.

Also, we refactored ztest a bit so it's easier to run an internal
test from the outside.