zq update through "refine semantic analaysis and fix bug with constants in filters" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2260, authored by @mccanne,
has been merged.

refine semantic analaysis and fix bug with constants in filters

This commit further develops the semantic analysis component of the
compiler.  This was needed to fix a problem where the lift-filter
optimization did not carry with it any Z constants that were defined
in the Z program causing such constant refs to be treated as field
references in the filters generated by the zbuf.Filter methods.

The fix is to fix the intermediate represenation of the Z program
to explicitly track the global scope of the Z program and initialize
constants in the runtime zctx before any filters are created.  Also,
when distributed workers are generated and use the AsProc() method
to zbuf.Filter, the AST returned here needs to include said constants
so this has been fixed.

In doing this refactor, we took one step toward the work needed
to split the parser AST and semantic DSL into logically distinct
components as outlined in issue brimsec/zq#2163.

In the new model, the compiler stages are all abstracted into a single
entry point in the compiler.Runtime object.  The ast command has been
reworked to show the AST at ay stage of compilation (after parsing,
after semantic analysys, after optimization, after parallelization,
and so forth).  Having the taps into compiler.Runtime from the ast
command makes it very easy to see what the various stages do and whether
they are working correctly and it allowed us to recast the complex
tests for the parallelization logic into simple ztests that use the
ast command.

In a future PR, we will more cleanly separate the parser AST from
the kernel DSL AST.  For now, the two logically separate AST forms
share the same data structure in compiler/ast.

Fixes brimsec/zq#2261