zq update through "Unflatten() function" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2277, authored by @henridf,
has been merged.

Unflatten() function

This PR adds an unflatten() function that turns successive dotted field names into nested records.  For example, unflattening `{"a.a": 1, "a.b": 1}` results in `{a:{a:1,b:1}}`.

Note that while unflattening is applied recursively from the top-level and applies to arbitrary-depth dotted names, it is not applied to dotted names that start at lower levels (for example {a:{"a.a":1}} is unchanged). This could be added but seems unnecessary given that the sole current use case for this function is zeek json records, where dottedness always starts at the root.

close brimsec/zq#2239