zq update through "Improvements to Grouping docs" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2288, authored by @philrz,
has been merged.

Improvements to Grouping docs

After seeing a community user invoke the familiar UNIX shell idiom `cut [field] | sort | uniq` in Z, I kept using it in my improvements I fed back to them. Someone on our team had to remind me to use just plain `by` (duh!), despite the fact that it was mentioned in the docs all along. To help myself not forget and maybe increase the chances community users will come to it on their own, here I've called out explicitly how the idiom _works_, but it's not advised.

While I was in there, I saw a TNZG->ZSON update worth making and a missing link.