zq update through "add intake backend resource" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2229, authored by @alfred-landrum,
has been merged.

add intake backend resource

This change adds a new backend resource type called an "intake", which can be configured with a Z shaper program and a target space. In this PR, I've create a single route (`POST /intake/{intake_id}/data`) that will run autodetection on the posted data, turning it into `*zng.Records`, and processing those records with a configured shaper, and then writing the resulting records to a space.

The intended usage is that intakes will eventually support different mechanism to receive or obtain data, with the next one planned to support data sent via the elasticsearch bulk api.

Closes brimsec/zq#2176