zq update through "Accept vars and aliases in shaper functions" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2289, authored by @henridf,
has been merged.

Accept vars and aliases in shaper functions

Previously the shaper functions required a literal type as second parameter, because they predated const and type literals. Now that those exist, the functions should also accept vars and type expressions. This commit does that.