zq update through "fluentbit to cloudwatch logging" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2299, authored by @mattnibs,
has been merged.

fluentbit to cloudwatch logging

Add helm chart responsible for installing fluentbit run as a daemonset in
a kubernetes cluster. The chart is currently configured to send logs to aws
CloudWatch. Once installed users can view in CloudWatch logs from the data
plane, applications and hosts.

The helm chart is derrived from the kubernetes template provided in this
tutorial:
https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/Container-Insights-setup-logs-FluentBit.html

Closes brimsec/zq#2274