zq update through "Change TZNG examples/refs to ZSON" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2312, authored by @philrz,
has been merged.

Change TZNG examples/refs to ZSON

Updating the user-facing docs updates in brimsec/zq#2294 made me aware of other lingering TZNG references in the docs that could also stand to be updated to ZSON. While I was in there, I found a few other links and missing pieces that seemed worth adding.

The zst spec was already pretty drafty so I wasn't 100% sure if the update to ZSON there was appropriate. But the fact I could actually use the comments in the ZSON to document the example in-line felt like enough of an improvement to justify it. I'm not sure it's enough to close the issues brimsec/zq#1714 and brimsec/zq#1734 (which appear to be duplicates of each other, FWIW) but since zst is his baby I'll wait to see what @mccanne has to say.