zq update through "add csv reader" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2317, authored by @mccanne,
has been merged.

add csv reader

This commit adds a reader for CSV formatted input.  It uses the
Go csv package, which conforms to RFC 4180.  CSV is not yet tied
into the auto-detector as it is currently too forgiving of a format
and accepts inputs that are clearly not intended as CSV.  See
issue brimsec/zq#2316  for more info.

Closes brimsec/zq#999.