zq update through "Temporal helm" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2306, authored by @marktwallace,
has been merged.

Temporal helm

The initial goal of the PR is to add a Temporal subchart to the Brim zservice “umbrella” Helm chart that we use for our zq-test cluster and automated tests. It also includes automated tests of the Temporal deployment and the zqd integration with Temporal.

Temporalio has developed a Helm chart that is relatively new (a few months old) compared to the overall project (which is years old). Their Helm chart is straightforward and easy to customize, but Temporal does not support its use as a subchart. So, in order to incorporate Temporal into our overall deployment, we needed to develop and test a pattern for using Temporal’s Helm chart as a subchart within the zservice chart.

One of the first things we needed was Helm repository from which we could reference `https://github.com/temporalio/helm-charts`. A Helm repository requires, at minimum, a URL for a tgz file and an index.yaml file with metadata for the chart and it dependencies. We created this within a new github repo, brimsec/helm-chart. This PR has a dependency on the new brimsec/helm-chart repo.

The procedure we used to create brimsec/helm-chart is documented in k8s/temporal.md. We added temporalio/helm-charts as a submodule of brimsec/helm-chart. We explicitly decided NOT to fork temporalio/helm-charts. We can and will fork temporalio/helm-charts if we need changes outside the capabilities of the current chart. Although temporalio/helm-charts follows some conventions that are different from ours (e.g. database password in environment variable) we will avoid forking it until we have to.

Temporal requires occasional database migrations. The procedures for their migrations are not automated. Because of this there is some manual setup for adding new test instances of Temporal. These are also documented in k8s/temporal.md. Without the detailed guide in k8s/temporal.md, the setup is difficult to get right, because there are multiple dependencies between steps. Temporal's complexity is hidden in the all-in-one docker image that we use for the ZTests in make test-system, but becomes more apparent the the K8s deployment that we use for the zq-test cluster and our future production environment.

Temporal has a concept of namespaces. Currently temporal namespaces require an external tool called `tctl` to set up. We may eventually be able to do this in our zqd code, but for now, follow the namespace setup instructions in k8s/temporal.md.

Minor changes to the zqd personality “temporal” were required to support K8s deployment. I think the temporal personality should have a reduced API, but I am not yet sure what we can remove.

The deployed footprint of the zservice chart is significantly larger after this PR. Seven new deployments and services have been added. (Use `kubectl get pod -A` to see the new pods in ci-master.)

A temporal cluster Ztest similar to the temporal system Ztest has been added to cd.yaml.

The new Makefile rules, test-temporal and helm-install-with-aurora-temporal depend on environment variables that have been added to cd.yaml and are described in k8s/temporal.md.