zq update through "add shape operator" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2327, authored by @mccanne,
has been merged.

add shape operator

This commit adds a Z operator to shape values of null:null into
their proper null type and coerces float64 to either uint64
or int64 if possible.  It does so by examining records that have
all the same field names where all numeric types are float64,
which is the case for unshaped csv and ndjson input.  It reads
the entire input, possibly spilling to disk, to determine the
proper types of everything then outputs the shaped values.
This is useful for cleaning up CSV inputs with various empty
values (i.e., null:null) appearing in various orders causes
the CSV reader to otherwise create many different record types
with varying null-typed columns.

We are generically calling this operator "shape" even though it's
functionality is initially limited as we expect this to be the seed
for an operator that will do many interesting stateful shaping operations.