zq update through "add default case syntax for Z-lang switch" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2336, authored by @mccanne,
has been merged.

add default case syntax for Z-lang switch

This commit adds a default case for a Z switch statement where
default is the same as a true match.  This means the default
statement must come last for it to work correctly.  We could
consider making the default case logically last, but I propose
we get this in so Phil can use it and we create an issue for
this if we want it.

Closes brimsec/zq#2322