zq update through "revert json encoding of nano.Duration" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2381, authored by @mccanne,
has been merged.

revert json encoding of nano.Duration

This commit reverts the way we marshal and unmarshal nano.Durations
so that we don't have to change the API used by the app whereever
a nano.Span was used.  It readopts the old format of {sec:float64,ns:float64}
to serialized durations.