zq update through "fix dot and add type values in zfmt" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2378, authored by @mccanne,
has been merged.

fix dot and add type values in zfmt

This commit fixes a bug in canonical zfmt to print "."  instead of
nothing for a root record field path.  It also adds support for
printing type values.

This addresses some but not all of the problems in brimsec/zq#2224.
We can add a number of tests for this stuff when we tackle
the rest of brimsec/zq#2224.