zq update through "add ast.RegExp" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2380, authored by @mccanne,
has been merged.

add ast.RegExp

This commit removes the regexp concept from the ast.Primitive path
since it is not a Z literal.  To do this, we added a RegExp operator
to the ast and runtime which simply is a boolean regexp pattern
matcher that operates on stringy types.

Now that Z primitives are cleaned up, we have a slate is clean
to add complex Z literals in a subsequent PR.