zq update through "preserve outer record type names in shaping and put ." by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2384, authored by @mccanne,
has been merged.

preserve outer record type names in shaping and put .

This commit fixes a bug where the type name of top-level records
was lost in shaping and in "put .".  The fix is to use zng.Type
instead of zng.TypeRecord where relevant and assert that said Type
is a TypeRecord optionally behind an alias.

Fixes brimsec/zq#2379