zq update through "Simplify Suricata shaper" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2400, authored by @philrz,
has been merged.

Simplify Suricata shaper

This is a change similar spirit to what I did with the Zeek shaper in brimsec/zq#2368. I tested this out by extracting the modified shaper from the Go and running it at the shell to confirm it creates ZSON that's byte-for-byte equivalent to the ones output from the prior shaper config as shown in the verification of brimsec/zq#1871 

```
$ zq -Z -I suricata-shaper-simplified.zed '| sort -r ts' eve.json > eve-shaped-simplified.zson

$ ls -l
-rw-r--r--    1 phil  staff   4948159547 Mar 23 07:49 eve-shaped-simplified.zson
-rw-r--r--    1 phil  staff   4948159547 Mar 22 13:32 eve-shaped.zson

$ diff eve-shaped-simplified.zson eve-shaped.zson
[no output]
```