zq update through "Add Parquet writer and rework Parquet reader" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2227, authored by @nwt,
has been merged.

Add Parquet writer and rework Parquet reader

\* Add zio/parquetio.Writer.
\* Add Parquet boomerang tests.
\* Rework zio/parquetio.Reader so boomerang tests pass.

This implementation uses github.com/brimsec/parquet-go, a fork of
github.com/fraugster/parquet-go with bug fixes and a larger set of
characters allowed in column names.

For simplicity, the new reader and writer use parquet-go interfaces that
require multiple memory allocations per record.  This is a step backward
from the current reader, but one that can be addressed later.

Closes brimsec/zq#2167.

Closes brimsec/zq#2263.