zq update through "allow record index by string as l-value" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2407, authored by @mccanne,
has been merged.

allow record index by string as l-value

This commit fixes a bug where l-values could not be index operations
with fields referenced as string, e.g., put .["x"] = 123.
The fix is to detect this case in the semantic pass and extend
the field-reference path with the string constant.

Fixes brimsec/zq#2326