zq update through "add ISO time literal" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2406, authored by @mccanne,
has been merged.

add ISO time literal

This commit adds support to the Z grammar for ISO time literals
and exposes durations and time literals into the expression
grammar.  We also added support for searches in the generic
expression compilation path as this can occur when match() is
used in expression context, e.g., put b=match(foo), which
we used for the time/duration literal testing.

Closes brimsec/zq#2335
Closes brimsec/zq#2195