zq update through "disable postpcap-suricata flaky test" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2415, authored by @mccanne,
has been merged.

disable postpcap-suricata flaky test

This test is flaky and many of us spend wasted emotion and time looking through the logs of a failed test only to discover postpcap-suricata was the culprit.  This test will be resurrected in `brimcap` so we can disable it here for now...