zq update through "add support for shaping to dynamic type expressions" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2426, authored by @mccanne,
has been merged.

add support for shaping to dynamic type expressions

This commit refactors the shaping logic into a const-type shaper
and a dynamic-type shaper.  In the latter case, the type expression
that is passed to a shaping function is evaluated for every input.
This allows the type expression to refer to the input record to
determine the type that should be applied, as is required in
an approach that shapes to a type based on a lookup into a map
value that might map a string value to the desired type for that
value.

Closes brimsec/zq#2321
Closes brimsec/zq#2383
Closes brimsec/zq#2390