zq update through "clean up zng.Record" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2437, authored by @mccanne,
has been merged.

clean up zng.Record

This commit cleans up zng.Record by recognizing that a zng.Record
is a special case of a zng.Value.  It replaces the type pointers and
Raw field with an embedded zng.Value.  This is the first step toward
improving the batch model as described in issue brimsec/zq#1762.

Closes brimsec/zq#1801