zq update through "import charm" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2447, authored by @mccanne,
has been merged.

import charm

This commit imports the charm package from github.com/mccanne.
This way we'll be able to work on it and extend it as discussed.

We also consolidated the uses golang.org/x/crypto/ssh/terminal
with a wrapper package in pkg/terminal (which is also used by
charm to get the interactive terminal width for text formatting).