zq update through "Update Zeek reference shaper to use a map" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2448, authored by @philrz,
has been merged.

Update Zeek reference shaper to use a map

This PR simplifies the reference shaper for Zeek by leveraging the brimsec/zq#2321 changes.

When I started this branch, I came to find that the previous shaper had stopped working due it containing words that became reserved as of the SQL changes in brimsec/zq#2392. Therefore I added the necessary escapes in this PR to make it work again. This made me pine a bit for the proposed brimsec/zq#2398 changes. I'll put some updates in that issue to remind us of the specifics of what failed here.

Since none of our automation had caught the shaper breakage, I've added a simple ztest here that would have caught it. Its goal is not to exhaustively test shaping for correctness, but just make sure it parses ok and can be minimally exercised.

To test that the generated output is equivalent for a large data set, here's side-by-side shaping of the [NDJSON zq-sample-data](https://github.com/brimsec/zq-sample-data/tree/main/zeek-ndjson) that give the same byte-for-byte output from both the older and new shaping configs, run through `zq` commit 7922fa6.

```
$ git branch
* main
  ref-shapers-use-maps

$ zq -Z -I shaper.zql '| sort ts,_path' ~/work/zq-sample-data/zeek-ndjson/* > /tmp/old.zson
```

```
$ git branch
  main
* ref-shapers-use-maps

$ zq -Z -I shaper.zed '| sort ts,_path' ~/work/zq-sample-data/zeek-ndjson/* > /tmp/new.zson
```

```
$ ls -l /tmp/*.zson
-rw-r--r--  1 phil  wheel  944433226 Mar 26 20:54 /tmp/new.zson
-rw-r--r--  1 phil  wheel  944433226 Mar 26 20:18 /tmp/old.zson

$ diff /tmp/old.zson /tmp/new.zson
[no output... files are the same]
```