Zed update through "zed query: remove -P flag" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2491, authored by @nwt,
has been merged.

zed query: remove -P flag

Instead, when join is the first operation in a Zed program, require two
input files, and use the first and second as the left and right join
inputs, respectively. Otherwise, merge the inputs by timestamp as usual.