Zed update through "extend len() to handle records" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2494, authored by @mccanne,
has been merged.

extend len() to handle records

This commit extends the len() function to return the number of
fields in a record when applied to a record value.  Previously,
len() raised an error when applied to record values.

Closes brimdata/zed#2374