Zed update through "add zson.FormatType()" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2511, authored by @mccanne,
has been merged.

add zson.FormatType()

This commits add the FormatType() function that formats up a
zng.Type in its canonical form.  The ZSON() method on zng.Type
didn't quite work right because it would output multiple
embedded typedefs when only the first is needed.  We may want
to get rid of the ZSON() method at some point as is commented
in issue brimdata/zed#2473.