Zed update through "zson: enable aliases for all primitive types" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2519, authored by @nwt,
has been merged.

zson: enable aliases for all primitive types

ZSON allows aliases for the integer, float, and stringy primitive types.
Extend that to all primitive types.

Note that the test omits a type value due to brimdata/zed#2518.