Zed update through "handle Zed bytes values in marshal/unmarshal" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2529, authored by @mccanne,
has been merged.

handle Zed bytes values in marshal/unmarshal

This commit changes the marshaling code to output a Zed bytes type
for Go byte slices and byte arrays and to detect when a bytes type
is being unmarshaled into a byte slice or array and unmarshal
accordingly.

We also fixed a bug where UnmarshalRecord failed when the out record
had a type name instead of a zng.TypeRecord.

Closes brimdata/zed#2475