Zed update through "split optimizer out of semantic analysis" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2558, authored by @mccanne,
has been merged.

split optimizer out of semantic analysis

Now that the AST and DAG representations are distinct, this
commit splits out the optimizer into its own package, cleanly
delineating its indepedence of the AST from the optimizer, as
it operates only on the DAG.  This sets us up to rewire the
merge logic in the optimizer so that a "from" Op can be added
more easily and dropped into the optimization framework
(e.g., it needs to be able to parallelize multiple sources on
different pools).