Zed update through "Move conn.log" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2574, authored by @philrz,
has been merged.

Move conn.log

I'm sure the `conn.log` has a limited time left in the Zed repo, but since it's currently referenced in the Zed README as being "located in this directory" I figured we should move it into there for now so the text becomes accurate. Or if we think its presence violates the purity of the new `cmd/zed/` directory, I'd be open to leaving it in its old location and just hyperlinking directly to it rather than saying "located in this directory". Just chime in with your preferences.