Zed update through "remove zio/ndjsonio.Reader in favor of zson.Reader" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2587, authored by @nwt,
has been merged.

remove zio/ndjsonio.Reader in favor of zson.Reader

Zed shaping renders the typed NDJSON reader unnecessary.  Replace it
with a zson.Reader and remove the following:

* api.LogPostRequest.JSONTypeeConfig (json_type_config request field
  for /space/{space}/logs/paths)
* json_config form field for /space/{space}/log
* -j flag to zed query and zq
* zio.ReaderOpts.JSON
* zio/ndjsonio.Reader
* zsontypes command

Closes brimdata/zed#2296.

`zeek/README.md` needs an update; I've left that for @philrz.