Zed update through "Mention JSON as a supported format in help text" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2608, authored by @philrz,
has been merged.

Mention JSON as a supported format in help text

I noticed the `zed query` help text doesn't mention the recently-added JSON support. While I was adding that, I also noticed that the following paragraph seemed to imply auto-detect was supported for all formats, which is not yet the case (brimdata/zed#2517). I briefly thought about getting into the minutiae of which formats are/aren't supported, but since the auto-detection provides guidance when it's given stuff it can't handle, I've opted to keep the text here simple.