Zed update through "Improvements to reference Zeek shaper" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2601, authored by @philrz,
has been merged.

Improvements to reference Zeek shaper

As a prerequisite to brimdata/zed#2594, here I'm proposing some changes to the reference Zeek shaper to make it less hazardous by renaming the `type id` to `type socket`. This avoids clashes when fields are referenced in the Zeek data that contain the common embedded record also known as `id`. Since we're now aware of the potential for clashes, brimdata/zed#2598 is open to track the improving of the error messaging to help make users aware when it happens.

While I was at it, I considered renaming `type port` as well, but my selection of the more descriptive `type networkport` led me to hit new bug brimdata/zed#2597. At that point I got to looking closer at the Zeek data and confirmed that there were no actual fields called `port`, which is perhaps unsurprising since that happens to be one of their native types, so perhaps they also have their own possible clashes to avoid. But while I've left `type port` intact, I did notice that there seemed be to be unnecessarily verbose specifications of `:port=(uint16)` throughout the various type definitions when a simple reference to the already-defined `:port` would have sufficed, so I've simplified those here as well.