Zed update through "fix bug that stopped help from being printed with no CLI args" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2618, authored by @mccanne,
has been merged.

fix bug that stopped help from being printed with no CLI args

This commit fixes a bug introduced when I blindly integrated feedback
in PR brimdata/zed#2596 :)  The logic here is restored to handle the case where
either the parse or the run result indicates that help should be displayed.

Fixes brimdata/zed#2607