Zed update through "add type name guard to grammar" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2612, authored by @mccanne,
has been merged.

add type name guard to grammar

This commit fixes a bug in the grammar where a type like "netx" would
match as "net" followed by x...  The fix is to add a guard to prevent this.

Closes brimdata/zed#2597