Zed update through "Fix link to RFC 8259" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2643, authored by @philrz,
has been merged.

Fix link to RFC 8259

I happened to try clicking the link in the ZSON spec and found it went to the wrong place. While I was fixing it I figured we might as well link directly to the anchor for the relevant section.

(Fun fact: @nwt actually "suggested" the correct link to the RFC in PR review comment [https://github.com/brimdata/zed/pull/1715#discussion_r533838382](https://github.com/brimdata/zed/pull/1715#discussion_r533838382) but then two characters were swapped in the commit that actually got merged, presumably because the edit was made outside of GitHub and then pushed. This illustrates the value of using the "accept suggestion" workflow so you can blame the person who made the suggestion if there's an error rather than making the mistake yourself. :trollface:)