Zed update through "simplify indexer and add explode operator" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2673, authored by @mccanne,
has been merged.

simplify indexer and add explode operator

This commit eliminates the "custom procs" from the indexing
logic and uses standard Zed queries instead.  This allowed us to
get rid of kernel.Hook and the logic for the custom compiler.

To support this, we converted and generalized the TypeSplitter
proc in the indexer as a new "explode" operator.  Explode
currently supports  just exploding by type but it is set up
to explode any container down the road.  We also just use
the standard cut proc instead of FieldCutter, which is slightly
less performant but the are other ways to skin the performance
cat.  Also, FieldCutter checks for uniform key types while
cut does not, but so does the index writer so problem solved.

With this stuff out of the way, it was also easy to remove
driver.Config as the only remaining vestige there was the zap.Logger.