Zed update through "report reasonable error when no zed lake root" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2701, authored by @mccanne,
has been merged.

report reasonable error when no zed lake root

This commit provides a better error when no lake root is specified.
While we were at it, we switch serivce.Config to use a storage.URI
instead of string to specify the root path.

Closes brimdata/zed#2671