Zed update through "change nameof() to return Zed missing instead of Go error" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2708, authored by @mccanne,
has been merged.

change nameof() to return Zed missing instead of Go error

In lieu of the more comprehensive work in issue brimdata/zed#1429, this commit
fixes the nameof() operator to return a proper Zed missing value.