Always populate location in tab history
The app state may contain history with an empty entries array.
If this happens, the location property will be undefined on the
history object. This will crash the app.

The fix is to never create a history with an empty initial entries.
If there are no entries, pass in undefined.

If you pass initial entries as undefined, the library will push a
"/" entry on the stack. Thus the location property will always be
populated.