Zed update through "Docs update post-zed/2692" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2744, authored by @philrz,
has been merged.

Docs update post-zed/2692

The docs changes in brimdata/zed#2692 got the automated examples in the markdown working (with the exception of the couple that were disabled), but some of the syntax in the nearby text still needed to be updated. I've taken care of that here, but while I was at it I also chipped away at several other things:

*   "ZQL" references have been changed to "Zed"
*   "Processors" has been renamed to "Operators", and because of this, I've tried to find ways to _not_ use the word "operator" in other contexts (e.g. saying "Boolean Logic" on the search page rather than "Boolean Operators")
*   Writing "records" in contexts that used to say "events" (which was a leftover from Zeek-centric speech)
*   Using Zed data types in places that used to reference Zeek data types (e.g. `addr` becomes `ip`)
*   Starting to describe searches and expressions as being usable anywhere in the pipeline
*   Starting to emphasize the importance of quoting strings on the RHS of field/value matches
*   Standardized on some formatting conventions, e.g.:
    *   How the quoted **Note:** blocks appear
    *   Paragraph blocks are wrapped at 80 characters where possible to make changes easier to review (exceptions being wide example outputs and tables)
*   I've run the pages through a spell/grammar checker

Things I've intentionally _not_ tried to attack in this iteration that will be addressed in future passes:

*   The examples still need to be rewritten with non-Zeek data
*   The sections on **Expressions** and **Data Types** are still woefully under-populated

Also, since there's inevitably things in the Brim changelog/wiki that link to the Zed language docs, once this PR is approved I'll create a Brim branch that points at this Zed branch and get that passing the link checkers in CI so I can merge that on the Brim side right after this on the Zed side.