Release notes bug (#1679)
* Turn on in dev and don't make ipc in reducer

An ipc call mistakenly made its way into the Launches reducer. The
reducer is run in both the main and the renderer process, because
it is global. When the ipc was called on the main process, there
was an error.

It would be good to also find a way to surface these errors more.

* Add an nvmrc file to make our node versions match those used in electron

* Log the error

* Test fix