Zed update through "Remove docs RE: non-working "Wildcard Field Names" syntax" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2773, authored by @philrz,
has been merged.

Remove docs RE: non-working "Wildcard Field Names" syntax

The `zq-command-disable` attached to these examples is a reflection of where we intentionally dropped support for `*=` and `in *`  as part of brimdata/zed#2692. Since a revived syntax to do the equivalent (& more) is not planned until a follow-on release (see brimdata/zed#2341 and brimdata/zed#2250), @mccanne and I agreed that we should just drop this from the docs for now, since we'd not be doing the `v0.25.0` user any favors by showing them syntax that won't work.