Zed update through "Add "zed lake init", change pool key spec, and .-to-this conversion" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2781, authored by @philrz,
has been merged.

Add "zed lake init", change pool key spec, and .-to-this conversion

Some users on public Slack expressed an interest in beta testing some of the new stuff, at which point they bumped into some out-of-date/missing stuff in the Zed lake design doc. I've addressed the first wave of speedbumps here by:

*   Adding coverage of `zed lake init`
*   Adapting to the new way pool keys are specified
*   Converting `.` references to `this`
*   Updates to `zed query` syntax now that pool specification and key range have moved into the Zed

I've also started linking to open GitHub Issues for the examples we show that aren't yet implemented. We've got stuff in the tooling that already tip the user off about this (e.g. if I execute `zed lake vacate` it responds with `"issue brimdata/zed#2545"`) but since we've got users trying to follow along with these examples in beta testing mode, I figure it'd be helpful to save them from trying to craft what they believe is the correct sequence of flags and commits only to then find out that it's not gonna work. If reviewers are on-board with this approach, I'll add notes-to-self in the linked issues with reminders to go back and update this doc as they get implemented.