Zed update through "rework TestMarkdownExamples to use bash" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2771, authored by @nwt,
has been merged.

rework TestMarkdownExamples to use bash

TestMarkdownExaamples can only run zq, and only against files in
zed-sample-data/zeek-default/. Rework it so zq-command blocks can
contain arbitrary commands executed by "bash -e -o pipefail" with a
working directory that defaults to the repository root but can be
specified with a word after zq-command in a block's info string.

We should remove zq from the names of lots of things touched here,
but that can happen later.