Zed update through ""make perf-compare" fixes" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2799, authored by @philrz,
has been merged.

"make perf-compare" fixes

This fixes a few problems I've noticed since I merged brimdata/zed#2785.

1.  After PR feedback, I'd modified the Workflow's `runs-on:` to point to `ubuntu-latest`, but the step that was downloading Zeek was still trying to install the version for Ubuntu 18.04, so it failed on its first regularly-scheduled run last night. The Zeek downloads are now for Ubuntu 20.04 and hence are in sync.
2.  When it failed last night, it ended up sending two notifications to Slack, as I'd forgotten we already have failures covered by the blanket `notify-main-failure.yaml`. So I've removed the redundant notification here.
3.  At that point when I re-tested I found two more failures waiting in line behind the now-fixed one for brimdata/zed#2783 that initially kicked off this exercise.
    1.  The perf script was not yet using `==` to test for equality in the Zed query, so that's now updated.
    2.  The problem described in https://github.com/brimdata/zed/pull/2123#issuecomment-859164320 was preventing the `-i ndjson -f zeek` permutation from running successfully, so the perf script now applies the reference Zed shaper for this case, and I've also added a bullet to the notes in the perf README to highlight this.

Since that all taught me that the perf script was way more out of sync than I thought, I went ahead and did a re-run on the appropriately-sized Google Cloud VM to update all the numbers at a current Zed commit. Some are a little faster, some are a little slower, but I also see that the `jq` ones are also slower, so, it might just be "hooray for public cloud variance".

In addition to the successful run on the cloud VM, I did make the updated Workflow run to completion in Actions, and the successful output is at https://github.com/brimdata/zed/runs/2797639874?check_suite_focus=true.