Zed update through "simplify and fix IDs in zng serializer" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2809, authored by @mccanne,
has been merged.

simplify and fix IDs in zng serializer

There were bugs in some of the zng serialization logic where
zng.Type.ID() was used instead of the zng.TypeID().  This commit
fixes these bugs and simplifies the code.  We also removed the
unused method zngio.Encoder.serializeTypes().