Zed update through "simplify Zed enums" by mccanne
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2820, authored by @mccanne,
has been merged.

simplify Zed enums

This commit simplifies Zed enums by removing the values from the
list of symbols.  The benefit of this was always dubious and avoiding
the crossing the type/value boundary (by encoding values in types)
simplifies things quite a bit.

Closes brimdata/zed#2508