Zed update through "Join doc" by philrz
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2836, authored by @philrz,
has been merged.

Join doc

Here's the long overdue doc for `join` as simple CLI-executable variations that use small sample data thanks to brimdata/zed#2627. I'll add some PR comments where I'm unsure if we're showing things in the best possible way and/or if there's a possible enhancement to file.

In addition to the examples I've shown here, there's one more I was hoping to include but I'm not able to come up with the correct syntax. I'm not sure if I'm just getting the syntax wrong or if it's not possible at all. My goal is to effectively show a "multi-way join" by using hierarchy in the dataflow to join two inputs first, then join a third input with the output of the first join. Following on from the last of the examples shown, let's say we had this additional data source that had the RGB values of the colors we use for all our fruit.

```
$ cat colors.ndjson
{"color":"red","hex":"#FF0000","rgb":{"r":255,"g":0,"b":0}}
{"color":"yellow","hex":"#FFFF00","rgb":{"r":255,"g":255,"b":0}}
{"color":"green","hex":"brimdata/zed#008000","rgb":{"r":0,"g":128,"b":0}}
{"color":"brown","hex":"#A52A2A","rgb":{"r":165,"g":42,"b":42}}
```

As separate operations, I can first join this with my `fruit.ndjson` records:

```
$ cat color-join.zed
from (
  file fruit.ndjson => sort color;
  file colors.ndjson => sort color;
) | inner join on color=color hex,rgb

$ zq -z -I color-join.zed | tee rgbfruit.zson
{name:"dates",color:"brown",flavor:"sweet",note:"in season",hex:"#A52A2A",rgb:{r:165,g:42,b:42}}
{name:"figs",color:"brown",flavor:"plain",hex:"#A52A2A",rgb:{r:165,g:42,b:42}}
{name:"avocado",color:"green",flavor:"savory",hex:"brimdata/zed#008000",rgb:{r:0,g:128,b:0}}
{name:"apple",color:"red",flavor:"tart",hex:"#FF0000",rgb:{r:255,g:0,b:0}}
{name:"strawberry",color:"red",flavor:"sweet",hex:"#FF0000",rgb:{r:255,g:0,b:0}}
{name:"banana",color:"yellow",flavor:"sweet",hex:"#FFFF00",rgb:{r:255,g:255,b:0}}
```

...then join _that_ output with my `people.ndjson`:

```
$ cat inner-join-rgb.zed
from (
  file rgbfruit.zson => sort flavor;
  file people.ndjson => sort likes;
) | inner join on flavor=likes eater:=name

$ zq -z -I inner-join-rgb.zed
{name:"figs",color:"brown",flavor:"plain",hex:"#A52A2A",rgb:{r:165,g:42,b:42},eater:"jessie"}
{name:"dates",color:"brown",flavor:"sweet",note:"in season",hex:"#A52A2A",rgb:{r:165,g:42,b:42},eater:"quinn"}
{name:"strawberry",color:"red",flavor:"sweet",hex:"#FF0000",rgb:{r:255,g:0,b:0},eater:"quinn"}
{name:"banana",color:"yellow",flavor:"sweet",hex:"#FFFF00",rgb:{r:255,g:255,b:0},eater:"quinn"}
{name:"apple",color:"red",flavor:"tart",hex:"#FF0000",rgb:{r:255,g:0,b:0},eater:"morgan"}
{name:"apple",color:"red",flavor:"tart",hex:"#FF0000",rgb:{r:255,g:0,b:0},eater:"chris"}
```

However, my attempt to come up with the correct syntax to do this all in one Zed script all become parse errors, such as:

```
$ cat multi-way-join.zed
from
(
  (
    file fruit.ndjson => sort color;
    file colors.ndjson => sort color;
  ) | inner join on color=color hex,rgb | sort flavor;
  (
   file people.ndjson => sort likes;
  )
) | inner join on flavor=likes eater:=name

$ zq -z -I multi-way-join.zed
zq: parse error: error parsing Z at line 4, col 23:
from
(
  (
    file fruit.ndjson => sort color;
                  === ^ ===
    file colors.ndjson => sort color;
  ) | inner join on color=color hex,rgb | sort flavor;
  (
   file people.ndjson => sort likes;
  )
) | inner join on flavor=likes eater:=name
```

Could someone who better understands the parser help me confirm if I'm barking up the wrong tree here? Thanks!