Provide explicit values to the pool search url
When no span params are provided, all the components will
implicitly use the pool defaults (saved in redux).
When new commits arrive, the
pool span will change. This results in
some of the UI re-rendering even though the url has not changed.

The only case we want this re-rendering behavior is when we are
dragging a pcap into the app. Then we want to have an empty span
and rely on the updating defaults. (although this feels confusing and
worthy of simplifying)

By being explicit with the spanArgs in the url, we can ensure that
the UI will not re-render when the span of the pool changes in redux.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch more-data-bug
# Changes to be committed:
#	modified:   app/summary/home.tsx
#