Zed update through "improve parse error formatting" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2864, authored by @nwt,
has been merged.

improve parse error formatting

When a parse error occurs, Zed tools print the full source code with no
source file information.  This is inhospitable when programs are long,
composed from files and the command line, or composed from multiple
files.  Improve this by printing only the source code line containing
the error along with a file name if appropriate.

Closes brimdata/zed#2391.