Copy html into dist
Previously, all html files were at the root of the project.
This commit adds a build step to copy each html file and place
it in the dist folder keeping its place in the folder structure.

This will afford more modular code where the html file can live
right next to the corresponding main process code that needs it.