HTML Dialog Module
This module mimics the electron native dialog module. It currently
has only one method: showErrorBox(title, content). The content
can be a markdown string.

It returns a promise and resolves when the dialog is closed.

A new icon was added for this, a new hook for keyboard shortcuts,
and a helper method to extract query parameters.