HTML Dialog Error Box for Windows Pre-25 Check (#1751)
* Copy html into dist

Previously, all html files were at the root of the project.
This commit adds a build step to copy each html file and place
it in the dist folder keeping its place in the folder structure.

This will afford more modular code where the html file can live
right next to the corresponding main process code that needs it.

* Create a core Markdown component

Anywhere we need to render markdown we can use this component.

* Secure browser windows before anything else runs

We dont want to open a browser window before these listeners are
setup. (We will do this in a subsequent commit)

* HTML Dialog Module

This module mimics the electron native dialog module. It currently
has only one method: showErrorBox(title, content). The content
can be a markdown string.

It returns a promise and resolves when the dialog is closed.

A new icon was added for this, a new hook for keyboard shortcuts,
and a helper method to extract query parameters.

* Use the htmlDialog in the windows pre-25 check

* Remove application menu on windows