Zed update through "replace fuse operator field renaming with union" by nwt
This is an auto-generated commit with a Zed dependency update. The Zed PR
https://github.com/brimdata/zed/pull/2885, authored by @nwt,
has been merged.

replace fuse operator field renaming with union

When records have a field with the same name but different types, the
fuse operator renames all but one uniquely by appending a suffix.
Remove this behavior and combine the different types in a union instead.

Closes brimdata/zed#2855.

Part of brimdata/zed#2852.

Depends on brimdata/zed#2881.